flight-app-py/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
